Add files via upload
The OSS Corporation is a globally distributed firm. They have their headquarters in East US with another branch office in SouthEast Asia. Currently, they are working on a project and decided that the application tier of this project will reside in one of its branch regions. For security reasons, OSS Corporation management is adamant on keeping their data tier in the headquarter region.

As an organization, they are open to suggestions and are currently evaluating Azure as a deployment platform. To prepare for deployment of IaaS Standard DS1 v2, OSS Corporation must deploy an IaaS v2 virtual network in the headquarter region for its database. But for the application, it should create another IaaS v2 virtual network in the branch region. In addition, because the communication between App and data should happen over a private channel, one needs to prepare their branch office virtual network for establishing connectivity to the headquarter’s IaaS v2 virtual network by creating a virtual network gateway and deploy a test IaaS Standard DS1 v2 VM to the virtual networks for verifying the connection. 

After the deployment team ensures the connectivity between both the networks, you can validate the same using Ping.